test(reference): consolidate GlobalMaxPool coverage
Keep the GlobalMaxPool regression cases in the existing reference evaluator test module instead of creating a dedicated file.
